auto generating guid and assign when creating a new project, renamed routes to views
