+++ app/form_validation.py
@@ -4,10 +4,7 @@ from wtforms.validators import InputRequired, Length, Regexp, NumberRange
 from datetime import date
 
 class CreateProject(FlaskForm):
-    guid = StringField('Project GUID', [ InputRequired(),
-        Regexp(r'[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}', message="Invalid GUID"),
-        Length(min=36, max=36, message="Invalid GUID length")
-        ])
+    guid = HiddenField()
     title = StringField('Project Title', [ InputRequired(),
         Regexp(r'^[A-ZÀ-ȕa-z0-9 ]*$', message="Invalid Project Title"),
         Length(min=3, max=75, message="Invalid Project Title length")
