+++ requirements.txt
@@ -11,3 +11,4 @@ cssmin
 Flask-WTF
 wtforms
 datetime
+fastapi_utils
